requirements.txt
django==2.2.5
djangorestframework==0.0.1
social-auth-app-django==3.1.0